🏗️ Architect: Modularized jovian_moons plotting
